🚚 api: Move models to session-management context
