+++ api/tests/tooling/domain-builder/factory/build-v3-certification-challenge-live-alert-for-administration.js
@@ -1,4 +1,4 @@
-import { V3CertificationChallengeLiveAlertForAdministration } from '../../../../src/certification/course/domain/models/V3CertificationChallengeLiveAlertForAdministration.js';
+import { V3CertificationChallengeLiveAlertForAdministration } from '../../../../src/certification/session-management/domain/models/V3CertificationChallengeLiveAlertForAdministration.js';
 
 const buildV3CertificationChallengeLiveAlertForAdministration = function ({ id = 456, issueReportSubcategory } = {}) {
   return new V3CertificationChallengeLiveAlertForAdministration({
